This should have fixed the panic errors from tess mapping.
Indeed: if we ask for 0 triangles / 0 discs, when trying to map the GPU
buffer, we end up with a `MapFailed` error from luminance, which is
great.

I think this bug is a very interesting issue that I should have in mind:
stop using the `unwrap()` function and write the code for both the
success and failure cases.